Setup basic home page
This uses a master template. A master template lays out the overall structure, headers, footers etc. It has section/areas which are sort of holes. They get filled in by the actual pages (that use the master as a tenmpate).

The home page then provides a section which fills in the hole in the master template